requirements.txt
contourpy==1.0.7
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
fonttools==4.39.4
future==0.18.3
kiwisolver==1.4.4
matplotlib==3.7.1
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pandastable==0.13.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tkterminal==0.1.0
tzdata==2023.3
xlrd==2.0.1
